move provider to pharmacySelect; fix formStore reset
