+++ packages/components/src/systems/PharmacySelect/index.tsx
@@ -1,1 +1,2 @@
 export * from './PharmacySelect';
+export * from './PharmacySelectionProvider';
