--- packages/components/src/systems/PharmacySelectionProvider/index.tsx
@@ -1,1 +1,0 @@
-export * from './PharmacySelectionProvider';
